revamped project for multiple users

+++ src/main/java/org/launchcode/patentinvestor/controllers/TagController.java
@@ -1,9 +1,7 @@
 package org.launchcode.patentinvestor.controllers;
 
 import org.launchcode.patentinvestor.data.TagRepository;
-import org.launchcode.patentinvestor.models.Comparators;
-import org.launchcode.patentinvestor.models.PaginatedListingService;
-import org.launchcode.patentinvestor.models.Tag;
+import org.launchcode.patentinvestor.models.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -13,6 +11,7 @@ import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
@@ -20,17 +19,29 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import static org.launchcode.patentinvestor.controllers.HomeController.NOT_LOGGED_IN_MSG;
+
 /**
  * Created by kamdem
  */
 @Controller
 @RequestMapping("tags")
-public class TagController extends AbstractBaseController {
+public class TagController {
+
+    //General messages
+    private final String INFO_MESSAGE_KEY = "message";
+
+    //About deleted tags from stocks
+    private final String ACTION_MESSAGE_KEY = "secondMessage";
 
     static final int numberOfItemsPerPage = 10;
     static final String baseColor = "white";
     static final String selectedColor = "green";
-    static String ordinarySortCriteria = "";
+
+    String tagSortCriteria = "";
+
+    @Autowired
+    AuthenticationController authenticationController;
 
     @Autowired
     private TagRepository tagRepository;
@@ -38,26 +49,49 @@ public class TagController extends AbstractBaseController {
     @Autowired
     private PaginatedListingService<Tag> paginatedListingService;
 
+    /**
+     * @param model
+     * @param page
+     * @param size
+     * @param sortIcon
+     * @param request
+     * @param redirectAttributes
+     * @return
+     */
     @RequestMapping(method = RequestMethod.GET)
     public String displayAllTags(
             Model model,
             @RequestParam("page") Optional<Integer> page,
             @RequestParam("size") Optional<Integer> size,
-            @RequestParam("sortIcon") Optional<String> sortIcon
+            @RequestParam("sortIcon") Optional<String> sortIcon,
+            HttpServletRequest request,
+            RedirectAttributes redirectAttributes
     ) {
+
+        User loggedInUser = authenticationController.getUserFromSession(request.getSession());
+        if (loggedInUser == null) {//user is NOT logged in
+            redirectAttributes.addFlashAttribute(INFO_MESSAGE_KEY, "danger|" +
+                    NOT_LOGGED_IN_MSG);
+            return "redirect:/";
+        } else {//user is logged in
+            model.addAttribute("isLoggedIn", true);
+        }
+
         int currentPage = page.orElse(1);
         int pageSize = size.orElse(numberOfItemsPerPage);
-        ordinarySortCriteria = sortIcon.orElse(ordinarySortCriteria);
+        tagSortCriteria = sortIcon.orElse(tagSortCriteria);
         String iconsDestinationUrl = "/tags/?sortIcon=";
         String paginationDestinationUrl = "/tags/?size=";
 
-        List<Tag> listOfTagsFound = (List<Tag>) tagRepository.findAll();
+        //List<Tag> listOfAllTags = (List<Tag>) tagRepository.findAll();
+        //Filter tags pertaining to this User's portfolio ONLY
+        List<Tag> listOfTagsFound = loggedInUser.getPortfolio().getTags();
 
         if (listOfTagsFound.size() == 0) {
             model.addAttribute(INFO_MESSAGE_KEY, "info|No investment fields were found! Click the 'create new field' button below to create one");
         }
 
-        switch (ordinarySortCriteria) {
+        switch (tagSortCriteria) {
             case "investmentFieldUp":
                 listOfTagsFound.sort(Comparators.tagComparator);
                 break;
@@ -69,11 +103,12 @@ public class TagController extends AbstractBaseController {
         paginatedListingService.setListOfItems(listOfTagsFound);
         Page<Tag> tagPage = paginatedListingService.findPaginated(PageRequest.of(currentPage - 1, pageSize));
 
-        model.addAttribute("title", "All investment fields");
+        model.addAttribute("title", "All investment fields of " +
+                loggedInUser.getUsername());
 
         //named stockPage because using the same fragment as that of stocks
         model.addAttribute("stockPage", tagPage);
-        model.addAttribute("sortIcon", ordinarySortCriteria);
+        model.addAttribute("sortIcon", tagSortCriteria);
         model.addAttribute("baseColor", baseColor);
         model.addAttribute("selectedColor", selectedColor);
         model.addAttribute("iconsDestinationUrl", iconsDestinationUrl);
@@ -91,24 +126,67 @@ public class TagController extends AbstractBaseController {
         return "tags/index";
     }
 
+    /**
+     * A Logged in User adds a tag to his own portfolio
+     *
+     * @param model
+     * @param request
+     * @return
+     */
     @GetMapping("create")
-    public String renderCreateTagForm(Model model) {
+    public String renderCreateTagForm(
+            Model model,
+            HttpServletRequest request,
+            RedirectAttributes redirectAttributes) {
+
+        User loggedInUser = authenticationController.getUserFromSession(request.getSession());
+        if (loggedInUser == null) {//user is NOT logged in
+            redirectAttributes.addFlashAttribute(INFO_MESSAGE_KEY, "danger|" +
+                    NOT_LOGGED_IN_MSG);
+            return "redirect:/";
+        }
+
         model.addAttribute("title", "Create investment field");
         model.addAttribute(INFO_MESSAGE_KEY, "info|'Investments fields' help categorize" +
                 " stocks. Keep the 'Field name' short, ideally a single word of less than 10 characters." +
                 " You may be more descriptive in the property 'Description of the investment field'.");
-        model.addAttribute(new Tag());
+
+        //Filter tags pertaining to this User's portfolio ONLY
+        //User loggedInUser = authenticationController.getUserFromSession(request.getSession());
+        model.addAttribute("isLoggedIn", true);
+        Portfolio portfolio = loggedInUser.getPortfolio();
+
+        model.addAttribute(new Tag(portfolio)); //model.addAttribute(new Tag());
         return "tags/create";
     }
 
+    /**
+     * A Logged in User adds a tag to his own portfolio
+     *
+     * @param tag
+     * @param errors
+     * @param model
+     * @param redirectAttributes
+     * @return
+     */
     @PostMapping("create")
     public String processCreateTagForm(
             @Valid @ModelAttribute Tag tag,
             Errors errors,
             Model model,
+            HttpServletRequest request,
             RedirectAttributes redirectAttributes) {
+
+        User loggedInUser = authenticationController.getUserFromSession(request.getSession());
+        if (loggedInUser == null) {//user is NOT logged in
+            redirectAttributes.addFlashAttribute(INFO_MESSAGE_KEY, "danger|" +
+                    NOT_LOGGED_IN_MSG);
+            return "redirect:/";
+        }
+
         if (errors.hasErrors()) {
-            model.addAttribute("title", "Create tag (investment field)");
+            model.addAttribute("title", "Create investment field");
+            model.addAttribute("isLoggedIn", true);
             model.addAttribute(tag);
             model.addAttribute(ACTION_MESSAGE_KEY, "danger|Failed to create a new investment field");
             return "tags/create";
@@ -118,28 +196,65 @@ public class TagController extends AbstractBaseController {
         return "redirect:/tags";
     }
 
+    /**
+     * A User should ONLY delete his own tags
+     *
+     * @param model
+     * @param request
+     * @return
+     */
     @GetMapping("delete")
-    public String displayDeleteTagForm(Model model) {
+    public String displayDeleteTagForm(
+            Model model,
+            HttpServletRequest request,
+            RedirectAttributes redirectAttributes) {
+
+        User loggedInUser = authenticationController.getUserFromSession(request.getSession());
+        if (loggedInUser == null) {//user is NOT logged in
+            redirectAttributes.addFlashAttribute(INFO_MESSAGE_KEY, "danger|" +
+                    NOT_LOGGED_IN_MSG);
+            return "redirect:/";
+        }
+
         model.addAttribute("title", "Delete investment field");
-        List<Tag> tags = new ArrayList<>();
-        Iterable<Tag> allTags = tagRepository.findAll();
+        List<Tag> deletableTags = new ArrayList<>();
+
+        //Filter tags pertaining to this User's portfolio ONLY
+        List<Tag> allTags = loggedInUser.getPortfolio().getTags();
+        //Iterable<Tag> allTags = tagRepository.findAll();
+
         for (Tag tag : allTags) {
             if (tag.getStocks().size() == 0) {
-                tags.add(tag);
+                deletableTags.add(tag);
             }
         }
-        model.addAttribute("tags", tags);
+        model.addAttribute("tags", deletableTags);
+        model.addAttribute("isLoggedIn", true);
         model.addAttribute(INFO_MESSAGE_KEY, "info|Only investment fields currently not set to any stock are listed. You may not delete any investment field currently tied to some stock");
         return "tags/delete";
     }
 
+    /**
+     * @param tagIds
+     * @param redirectAttributes
+     * @return
+     */
     @PostMapping("delete")
     public String processDeleteTagForm(
             @RequestParam(required = false) int[] tagIds,
+            HttpServletRequest request,
             RedirectAttributes redirectAttributes) {
+
+        User loggedInUser = authenticationController.getUserFromSession(request.getSession());
+        if (loggedInUser == null) {//user is NOT logged in
+            redirectAttributes.addFlashAttribute(INFO_MESSAGE_KEY, "danger|" +
+                    NOT_LOGGED_IN_MSG);
+            return "redirect:/";
+        }
+
         if (tagIds != null) {
             for (int id : tagIds) {
-                //tagRepository
+                Tag tag = tagRepository.findById(id).get();
                 tagRepository.deleteById(id);
             }
             redirectAttributes.addFlashAttribute(ACTION_MESSAGE_KEY, "success|" + tagIds.length + " investment field(s) deleted");
@@ -167,10 +282,27 @@ public class TagController extends AbstractBaseController {
         return "tags/view";
     }*/
 
+    /**
+     * @param model
+     * @param tagId
+     * @return
+     */
     @GetMapping("edit/{tagId}")
     public String displayEditForm(
             Model model,
-            @PathVariable int tagId) {
+            @PathVariable int tagId,
+            HttpServletRequest request,
+            RedirectAttributes redirectAttributes) {
+
+        User loggedInUser = authenticationController.getUserFromSession(request.getSession());
+        if (loggedInUser == null) {//user is NOT logged in
+            redirectAttributes.addFlashAttribute(INFO_MESSAGE_KEY, "danger|" +
+                    NOT_LOGGED_IN_MSG);
+            return "redirect:/";
+        } else {//user is logged in
+            model.addAttribute("isLoggedIn", true);
+        }
+
         Tag tag = tagRepository.findById(tagId).get();
         String tagName = tag.getDisplayName();
         model.addAttribute("title",
@@ -185,6 +317,13 @@ public class TagController extends AbstractBaseController {
         return "tags/edit";
     }
 
+    /**
+     * @param tagId
+     * @param name
+     * @param description
+     * @param redirectAttributes
+     * @return
+     */
     @PostMapping("edit")
     public String processEditForm(
             int tagId,
